fixed more compiler warning

+++ include/table.h
@@ -115,13 +115,13 @@ inline Table::Table(Table &tbl)
 inline ostream &operator<<(ostream &out, const Table& t)
 {
     if ( t._has_yerr ) {
-        for(int i=0; i<t._x.size(); ++i) {
+        for(size_t i=0; i<t._x.size(); ++i) {
             out << t._x[i] << " " << t._y[i] << " " << t._yerr[i] << " " << t._flags[i] << endl;
         }
     }
     else {
         //out << t.size() << endl;
-        for(int i=0; i<t._x.size(); ++i) {
+        for(size_t i=0; i<t._x.size(); ++i) {
             out << t._x[i] << " " << t._y[i] << " " << t._flags[i] << endl;
         }
     }
